Changed damagedealer to obstacle spawner
