+++ ContextDesktop/src/main/java/com/github/migi_1/Context/damageDealers/EnemySpawner.java
@@ -1,5 +1,6 @@
 package com.github.migi_1.Context.damageDealers;
 
-public class EnemySpawner {
 
+public class EnemySpawner {
+   
 }
